[IMPORTANT UPDATE] Done logs!

+++ src/main/java/org/eml/fox/files/PropertiesUtils.java
@@ -1,10 +1,14 @@
 package org.eml.fox.files;
 
 
+import org.apache.commons.configuration2.PropertiesConfiguration;
+import org.apache.commons.configuration2.PropertiesConfigurationLayout;
+import org.apache.commons.configuration2.ex.ConfigurationException;
 import org.eml.fox.Main;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.*;
+import java.nio.charset.StandardCharsets;
 import java.util.Objects;
 import java.util.Properties;
 
@@ -26,7 +30,8 @@ public class PropertiesUtils {
      * @param target
      * @return
      */
-    public static void copy(@NotNull File source, File target) {
+    public static void copy(@NotNull File source, @NotNull File target) {
+        if (target.exists())return;
         if (source.isFile()) {
             copyFile(source, target);
         } else {
@@ -79,11 +84,15 @@ public class PropertiesUtils {
     }
 
 
+    /**
+     * @param key key
+     * @param value value
+     */
     public static void writeToProperties(String key , String value) {
         Properties properties = new Properties();
         try {
             FileOutputStream fos = new FileOutputStream(Main.USER_DIR+"\\user.properties", true);
-            OutputStreamWriter opw = new OutputStreamWriter(fos, "utf-8");
+            OutputStreamWriter opw = new OutputStreamWriter(fos, StandardCharsets.UTF_8);
             properties.setProperty(key, value);
             properties.store(opw , null);
             fos.close();
@@ -93,5 +102,43 @@ public class PropertiesUtils {
             e.printStackTrace();
         }
 
+    }
+
+    /**
+     * @param propertiesName name
+     * @param key key
+     * @param value value
+     */
+    public static void updateProperties(String propertiesName,String key,String value){
+        String path = getDir();
+        String filePath = path + "\\" + propertiesName;
+        File file = new File(filePath);
+        PropertiesConfiguration configuration = new PropertiesConfiguration();
+        PropertiesConfigurationLayout layout = configuration.getLayout();
+        try {
+            layout.load(configuration,new InputStreamReader(new FileInputStream(file)));
+            configuration.setProperty(key,value);
+            layout.save(configuration,new OutputStreamWriter(new FileOutputStream(file)));
+        }catch (IOException | ConfigurationException e){
+            e.printStackTrace();
+        }
+    }
+
+    /**
+     * @param propertiesSRCName p
+     * @param key k
+     * @param value v
+     */
+    public static void updateSRCProperties(String propertiesSRCName,String key,String value){
+        File file = new File("src/main/resources/"+propertiesSRCName);
+        PropertiesConfiguration configuration = new PropertiesConfiguration();
+        PropertiesConfigurationLayout layout = configuration.getLayout();
+        try {
+            layout.load(configuration,new InputStreamReader(new FileInputStream(file)));
+            configuration.setProperty(key,value);
+            layout.save(configuration,new OutputStreamWriter(new FileOutputStream(file)));
+        }catch (IOException | ConfigurationException e){
+            e.printStackTrace();
+        }
     }
 }
